+++ src/main/java/com/kopieczek/gamble/savefiles/SaveFileDb.java
@@ -3,6 +3,6 @@ package com.kopieczek.gamble.savefiles;
 import java.util.Optional;
 
 public interface SaveFileDb<T> {
-    void put(T key, int[] ramData);
-    Optional<int[]> get(T key);
+    void put(T key, byte[] ramData);
+    Optional<byte[]> get(T key);
 }
